server/requirements.txt
python-socketio==5.5.1
starlette==0.26.1
uvicorn[standard]==0.21.1